Context::Net Initial Commit
Context::Net is a simple socket wrapper based heavily on eggdrop2.0
libeggdrop network code.  It's sole idea is to be small/simple code
to be dropped into a project directly vs used as a library.  I dont
like making projects that depend on a dozen libraries so this is a
small peice that is versatile and simple to use for any project that
needs to handle a lot of non-blocking sockets.

It is NOT complete, though currently compiles on OS X, FreeBSD (6.1),
and Linux (2.6, glibc 2.3), and will stay that way.  It will later
support epoll and kqueue on systems that support it.  It should
compile on any operating system supporting poll and vasprintf.